+++ headers/simdbinarypacking.h
@@ -43,7 +43,7 @@ public:
             size_t &nvalue) {
         checkifdivisibleby(length, BlockSize);
         const uint32_t * const initout(out);
-        *out++ = length;
+        *out++ = static_cast<uint32_t>(length);
         while(needPaddingTo128Bits(out)) *out++ = CookiePadder;
         uint32_t Bs[HowManyMiniBlocks];
         for (const uint32_t * const final = in + length; in + BlockSize
@@ -117,7 +117,7 @@ public:
             size_t &nvalue) {
         checkifdivisibleby(length, BlockSize);
         const uint32_t * const initout(out);
-        *out++ = length;
+        *out++ = static_cast<uint32_t>(length);
         uint32_t Bs = maxbits(in,in + length);
         *out++ = Bs;
         while(needPaddingTo128Bits(out)) *out++ = CookiePadder;
